부산대 BE_김해중 2단계 - 관리자 화면 (#104)
* feat(view): Thymeleaf 연동

* feat(view): 상품 조회 구현

* feat(view): 상품 수정 구현

* refactor(view): 단어 변경 (edit->update)

* feat(view): 상품 추가 구현

* feat(view): 상품 삭제 구현

* refactor(view): 타임리프 JS 분리

* refactor(view): CSS 분리 및 테이블,모달 디자인 개선

* docs(readme): 관리자 화면 관련 구현 기능 목록 정리

* refactor(view): DTO 내 id 제거 및 뷰 컨트롤러 수정

* fix(view): 너무 큰 price값 입력 시 발생하는 에러 수정

+++ src/main/java/gift/dto/ProductRequestDTO.java
@@ -1,21 +1,22 @@
 package gift.dto;
 
+import java.math.BigInteger;
+
 public class ProductRequestDTO {
     private final String name;
-    private final Integer price;
+    private final BigInteger price;
     private final String imageUrl;
 
-    public ProductRequestDTO(String name, Integer price, String imageUrl) {
+    public ProductRequestDTO(String name, BigInteger price, String imageUrl) {
         this.name = name;
         this.price = price;
         this.imageUrl = imageUrl;
     }
+
     public String getName() {
         return name;
     }
-    public Integer getPrice() {
-        return price;
-    }
+    public BigInteger getPrice() { return price; }
     public String getImageUrl() {
         return imageUrl;
     }
