+++ src/main/java/gift/dto/ProductResponseDTO.java
@@ -1,12 +1,14 @@
 package gift.dto;
 
+import java.math.BigInteger;
+
 public class ProductResponseDTO {
     private final Integer id;
     private final String name;
-    private final Integer price;
+    private final BigInteger price;
     private final String imageUrl;
 
-    public ProductResponseDTO(Integer id, String name, Integer price, String imageUrl) {
+    public ProductResponseDTO(Integer id, String name, BigInteger price, String imageUrl) {
         this.id = id;
         this.name = name;
         this.price = price;
@@ -18,7 +20,7 @@ public class ProductResponseDTO {
     public String getName() {
         return name;
     }
-    public Integer getPrice() {
+    public BigInteger getPrice() {
         return price;
     }
     public String getImageUrl() {
